Added 'VersionCheck.java' and implemented it.
Implemented basic version checking against a github raw url found in the src folder called 'version' it will change as the plugin version does to ensure compatibility.

Gives a possibility to create a custom batch file to change the file name along with building the jar for the usage in a test server.